remove main from shared library
